Rename leveldb.rs -> binding.rs
