+++ src/database/cache.rs
@@ -1,5 +1,5 @@
 //! Structs and traits to work with the leveldb cache.
-use crate::leveldb::{leveldb_cache_create_lru, leveldb_cache_destroy, leveldb_cache_t};
+use crate::binding::{leveldb_cache_create_lru, leveldb_cache_destroy, leveldb_cache_t};
 use libc::size_t;
 
 #[allow(missing_docs)]
