+++ src/database/compaction.rs
@@ -1,7 +1,7 @@
 //! Compaction
 use super::key::Key;
 use super::Database;
-use crate::leveldb::leveldb_compact_range;
+use crate::binding::leveldb_compact_range;
 use libc::{c_char, size_t};
 
 pub trait Compaction<'a, K: Key + 'a> {
