+++ src/lib.rs
@@ -56,11 +56,11 @@ pub use crate::database::kv;
 pub use crate::database::management;
 pub use crate::database::options;
 pub use crate::database::snapshots;
-pub use crate::leveldb::{leveldb_major_version, leveldb_minor_version};
+pub use crate::binding::{leveldb_major_version, leveldb_minor_version};
 
 #[allow(missing_docs)]
 pub mod database;
-pub mod leveldb;
+mod binding;
 
 /// Library version information
 ///
