+++ src/mod.rs
@@ -1,2 +1,2 @@
 pub mod database;
-pub mod leveldb;
+mod binding;
